Break dependency to satisfy `cargo publish`.
Cargo won't publish any crate in the same workspace with a new version if another crate in the same version depends on that very version of the crate, which is not published yet. So we need to artificially temporarily break the dependencies of the other crate to allow publishing its dependency.